Dispatcher events could be lost when event channel is full (#48)
When the channel is full start dispatching events in a non-blocking manner. Events are queued and will time out after the dispatchTimeout.
The number of queued events is limited to the larger value of 10,000 .or 10% of the eventChannelCapacity. When the wait queue exceeds that limit the system will panic.

A new flag is added called dispatchTimeout to allow setting the timeout in seconds for dispatched events to timeout (DefaultDispatchTimeout = 300sec). If an event times out the event will still be lost.