Perf improvement: refine scheduling logic in shim to reduce num of events (#57)
This patch refines the main scheduling logic in shim layer in order to improve the performance. It removes most of the events from the main scheduling interval, because otherwise, it may create too may duplicate events that potentially block the dispatcher. This fix still ensures the consistency of the states in the state-machine.