Fix endless application recovery (#63)
Application recovery could end up in an endless loop waiting for all application to change state to Accepted. The application state transition was optimised to remove an unneeded wait and event. This optimisation (#57) caused the recovery to get stuck waiting for an app to reach the Accepted state while it already passed that state. Removing the call back that caused this issue.

This change also removes the obsolete test TestApplicationSchedulingWithoutDuplicateEvents