Add binding of volumes to the scheduler (#68)
In certain cases, even if the PVC and PV are defined correctly on the pod, the volume were not bound correctly when the pod was allocated. Added AssumePodVolumes and BindPodVolumes to the logic to make sure that all types of volumes are bound correctly. The bind state of the volumes is kept in the cache as part of the assumed pod info. Task function signatures were updated to align with the change
Predicate init was cleaned up to add further volume predicates

Co-authored-by: Wilfred Spiegelenburg <wspiegelenburg@cloudera.com>