Pod scheduling fixes and improvements (#73)
Before passing a pod to the scheduler volumes are sanity checked to prevent predicate failures later in the process. 
K8s auto scaling expect pods in an unschedulable state to consider auto scaling. The shim now automatically sets that state if not scheduled after a time out.
Application ID generation is modified to generate a simpler, more compact ID to fit in with other admission controllers. The image pull in the helm-chart has to changed to always for auto update support.