Add ability to support 3rd party K8s operator integration (#72)
K8s operators add the ability to provide details on applications for scheduling in K8s. They rely on CRDs for integration. This feature add the framework to support operators. The generic label based scheduling is moved to the general operator and the spark operator is added as the spark-k8s-operator.
The full design is documented in available the core documentation.