[YUNIKORN-45] Release task from pending state (#87)
Tasks are not released when they are removed while in a pending state.
This causes a leak on the core scheduler side due to asks not being
cleaned up. The core will schedule the ask and turn it to an allocation.
In the end this can lead to the core thinking that a node is out of
resources without traceable usage on the K8 side.