[YUNIKORN-87] Add predicate check for reservation
Add a flag to distinguish between allocation and reservation for
predicate checks. A reservation predicate run will not update the event
system in k8s. A new limited ist of predicates to run for reservations
only is created. The list of predicates for reservations must be
contained in the default and or configured list.