[YUNIKORN-176] Handle node create from pod (#123)
A pod can be created with a node set. If the node is not known yet a new
node is created to allow adding the pod. When the node event is received
the node that was added for the pod processing is not updated correctly.
The node and pod are not linked which causes the predicates to fail.

If this pod is a daemon pod the node will not become available for
scheduling.

Fixes: #123