[YUNIKORN-200] Add namespace quota annotations (#132)
Namespace quotas are enforced on submit of a pod. This prevents a pod
from being submitted if the quota is reached and does not allow for
queing of pods and scheduling when quota becomes available..
To support queing replace namespace quota with an annotation on the
namespace. The annotation is then passed on to the core with every new
application and set on the queue. Note that this functionality requires
the use of tag (namespace) based placement rules as the quota is not
modified on confiured queues in the config.

Annotations to be set:
yunikorn.apache.org/namespace.max.cpu
yunikorn.apache.org/namespace.max.memory

Fixes: #132