[YUNIKORN-251] use pod UID for recovered allocations (#140)
A task is created the allocation key is set to the pod UID. When the pod
is recovered the allocation key is set to the pod name. This is not
consistent and causes 2 issues:
- the allocation key might become empty when the name is empty
- the behaviour is not consistent for new and recovered pods

A side effect of the empty name is that if the allocation is deleted the
shim sends an empty ID to the core which removes all allocations instead
of just the requested one.

Fixes: #140