[YUNIKORN-241] fix the pod ordering in k8shim (#145)
Pod are converted into tasks for an application. The tasks are stored in
a map and returned as a slice. This causes tasks to always be returned
in a random order. This will affect the order in which they are send to
the core for scheduling.
The change adds a create timestamp to a task. The create timestamp is
the point in time the pod is recieved from K8s. This might differ from
the original submit time to K8s as it is an async method.

Tasks when fetched from the cache are always sorted based on their
create timestamp, oldest to youngest.

Fixes: #145