[YUNIKORN-231] update pod status based on feedback (#149)
Support for auto-scaling was implemented with a hard coded pod status
update after 3 seconds. This caused over provisioning of nodes.
This changes removes the hard coded update and uses the core state
feedback to mark pods correctly.

The logic for updates is driven by the core as it knows the correct
scheduling state of the pod. This just implments the forwarding of the
state to the k8s objects.

Fixes: #149