[YUNIKORN-282] stale nodes removal from cache (#158)
Nodes can be removed and added back again using the same identifier.
The informer can change that remove and add into an update. This can
cause stale information in the shim cache.
The fix makes sure the cache is correctly cleaned up in this case.
YUNIKORN-220 relies on the cache to be clean.

Fixes: #158