[YUNIKORN-268] Implement shim side remove application API (#155)
Previously the shim side remove app API was not fully implemented, this PR completes
that by sending the remove app request to the core, to make sure shim and core side state
to be sync'd. The remove API will fail fast if the shim finds the app still has non-terminated
tasks, we do not handle the complexity of how to cleanup pods today.