[YUNIKORN-291] Skip updatePodCondition in tests (#162)
After the changes for YUNIKORN-231 the core scheduler can push updates
for pods after 1 second and mark them unschedulable.
Since unit test does not mock the scheduler and there is no cluster some
of the structures needed in the updates will be nil. This causes the
tests to crash if they run for longer than 1 second. The full fix is to
mock the scheduler but that is a larger change.

Fixes: #162