[YUNIKORN-376] Update K8s dependency to 1.16 (#184)
Update the version of K8s we build against to 1.16.13, currently the
latest in this minor release. This is the first version of K8s that uses
go.mod for YuniKorn. The go.mod file is kept as clean as possible using
replace directives.

Predicates have been removed as per the removal in K8s itself.

The application CRD is regenerated against the new K8s version.

Fixes: #184