[YUNIKORN-366] Add rest API to update queue configuration (#182)
This PR adds a flag "enableConfigHotRefresh" in the scheduler startup option
when this flag is turned off, which disables the config hot-refresh. In this case, the user
can use the REST API to achieve the config update. When the config update is through
the REST, the updated configuration will be saved in the configmap.