[YUNIKORN-449] Added MatchNodeSelectorPred to reservation predicates list (#202)
This PR adds MatchNodeSelectorPred to the reserving phase predicate function list,
this makes sure while doing the reservation, the node-selector is honored in order to
increase the availability of each reservation.  As a side of the implementation, this PR
splits the previous GeneralPred to PodFitsResourcesPred, HostNamePred, this is more
flexible to register all the predicates and run a different set of predicates for different phases.


PodFitsHostPortsPred and MatchNodeSelectorPred.