[YUNIKORN-549] app recovery depends on pod order (#235)
Application recovery is currently linked to the pod order. This can
cause issues when the first pod of an application is a pod that was not
scheduled before (pending). The application would be marked as not
needing recovery even if all other pods of the application were
scheduled before (allocated).
Change the logic to make sure the application is marked correctly as
recovering independent of the pod list order.

Fixes: #235