[YUNIKORN-562] Support add labels/annotations to taskGroup (#239)
Sometimes we need to pass in the labels/annotations to the placeholder pods in order to make them more "like" the real pods.
This helps certain cases, for example, if some namespace has sidecar container auto-injection enabled, we can leverage the annotation field to enable or disable the injection,
just to comply with the real pod definition.

This PR adds labels/annotations to the taskGroup definition and makes it configurable, so the client can pass in the required values if needed.

Fixes: #239