[YUNIKORN-584] App recovery is skipped when applicationID is not set in pods' label. (#246)
When we list applications, we have a filter that checks label applicationId exists in pods, pods without this label will be excluded in this scan.
When we recover Spark jobs, spark pods do not necessarily have this label, we can recognize them using spark-app-selector. This caused the job to get excluded from the recovery app list, and then subsequent node recovery also failed because it finds the existing allocation for the job but the job was not recovered in the scheduler.

Fixes: #246