[YUNIKORN-600] PlaceholderManager structure init
Fix the initialisation of the PlaceholderManager structure to init the
channel and orphanPods map (renamed).
Fix the tests to call NewPlaceholderManager to not init the structure in
the test code and thus hiding init issues.

The stop channel is never closed and should be reused. We should not
create a new channel in the Start but in the New call. Fix up the Stop
checks to not use one long 5 second timeout but check every 100ms. This
change speeds up the test (down to 0.2s from 8s) and the shutdown of the
whole shim.