YUNIKORN-679: Add gangSchedulingStyle scheduling parameter (#269)
Introduced soft/hard gang scheduling styles, each app can specify if they want to use soft or hard
gang scheduling styles. The default style is soft, what means that after the placeholder timeout
the application will be scheduled as a simplee appliction without a gang request.
This PR also adds a Resuming state to represent the phase when the app fails to reserve resources
and starts to fall back to the normal scheduling cycle.

Fixes: #269