[YUNIKORN-741] Regression: occupied resources calculation (#279)
A pod that is scheduled by a different scheduler could be counted twice
as part of the occupied resources. This only occurs if the pod during
recovery was marked as pending and then shortly after recovery moves to
running.
Recovery is doing the correct thing the change is part of the normal
processing to limit the instances that the occupied resources are
updated.

This regression was introduced by YUNIKORN-677

Fixes: #279