[YUNIKORN-923] Application stateawre disable tag (#323)
If an application is submitted without an application ID the admission
controller generates a new ID. In this case there is no knowledge around
the number of allocations for the application and we should assume there
will not be any follow up pods.
In the case the application end up in a queue that has a state aware
pollicy it should not block waiting for the next allocation. If the
admission controller generates an application ID is also sets the
disable stateaware tag.
The shim propagates the tag onto the application object and passes it
onto the core for handling.

Fixes: #323