[YUNIKORN-508] fix log collection path (#329)
The e2e tests generate a log for analysis. The log directory is created
as part of the test run. The default path set in the code must be a
creatable or pre-existing path.

If the directory exists it is re-used. Files inside the directory will
be truncated and re-written when needed.
Changing default to: /tmp/e2e-test-reports

Closes: #329