[YUNIKORN-1124] Always pass initialised Attributes map (#390)
The core expects at least an empty map for the attributes on a node on
each request. The expectation is based on the fact that the partition
name should be sent. The default partition however can be omitted and
thus the map can be nil.
Defensive change and always sent an empty map.

Closes: #390

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>