[YUNIKORN-1243] Build replacement web UI for e2e tests (#439)
Build a simple web UI replacement image so we no longer rely on a web
image with the latest tag in the public hub. The web image runs as a
non root user and can run as a dynamic user on RedHat Openshift.

Includes major cleanup of the run_e2e_tests.sh script.

Closes: #439

Signed-off-by: Craig Condit <ccondit@apache.org>