[YUNIKORN-1320] [shim] Handle new userInfo annotation (#475)
Process the new annotiation from a pod that supports users and groups.
If no user is found the default user is returned, groups can be an empty
list and in that case a nil is returned.

If no annotations are found the fallback is to the deprecated user
label which only supports reading a username.

Closes: #475

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>