[YUNIKORN-1597] Recover assigned pods in Bound state (#541)
A pod that is recovered already assigned to a node is recovered with the
state set to Allocated. A recovered pod will not change state in that
case and wil stay as Allocated until it finishes.

If that pod is a placeholder it will not be included in the calculation
for the application state. Only Bound pods are counted for triggering
the application state change.
With gang scheduling this would leave the application with less than the
required placeholders in the right state to progress to the next state.
Never releasing the originator pod and thus breaking progress.

Closes: #541

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>