[YUNIKORN-1647] Dynamic namespace support in admission controller (#557)
Support annotations on a namespace to define processing behaviour in the
admission controller. The annotations have preference over the regexp.
If no annotations are found fall back on the regexp. If an annotation is
found it defines the behaviour and overrides what is defined in the
regexp.

Annotations to be supported on a namespace:
* yunikorn.apache.org/namespace.generateAppId
* yunikorn.apache.org/namespace.enableYuniKorn
Supported values for the annotation: true and false.

Closes: #557

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>