[YUNIKORN-1670] Make application recovery more robust
Remove timeout from application recovery, and ensure that applications
are removed from the recovery list as soon as they have transitioned
out of New or Recovering status (rather than into Accepted). This
prevents failure on large clusters due to timeouts, and also prevents
failure in the case of a rejected application.