[YUNIKORN-1707] Ignore pods in non-labeled namespaces (#579)
The admission controller allows pods in configured namespaces to be either
bypassed entirely, or passed through to YuniKorn but not labeled for
processing (useful in the plugin deployment to pass through to the embedded
default scheduler). However, if YuniKorn detects an applicationId or Spark
spark-app-selector, it will attempt to schedule that Pod via YuniKorn
regardless. This may not succeed if there is not a valid queue available
which matches placement rules.

To prevent this, the admission controller will now annotate pods which are
in non-labeled namespaces to mark them as ignored by YuniKorn, and the
scheduler will honor this annotation. This ensures that a namespace is
either processed completely by YuniKorn, or not at all.

Closes: #579