[YUNIKORN-1273] Option to generate unique app IDs (#594)
An option is added to the admission controller to control the generation
of the applicationID. By default all the pods in a namspeace that do not
have an application ID set will be grouped together as one application.

The option "admissionController.filtering.generateUniqueAppID", by
default false, when set to true will generate unique application IDs for
each pod.

Closes: #594

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>