[YUNIKORN-1908] Support GPU in placeholders
Extended resources and huge pages can not over commit. The pod validator
checks for a Limit setting in the pod spec. The pod fails validation
without it. Placeholder creation must make sure that the pods pass the
validator.
Set limits for all extended resources and hugepages-* as they cannot be
over committed.

Clean up of duplicated task group definition in the test.