[YUNIKORN-1908] Support GPU in placeholders (#651)
Extended resources and huge pages can not over commit. The pod validator
checks for a Limit setting in the pod spec. The pod fails validation
without it. Placeholder creation must make sure that the pods pass the
validator. Set limits for all extended resources and hugepages-* as they
cannot be over committed.

Closes: #651

Signed-off-by: Craig Condit <ccondit@apache.org>