[YUNIKORN-2366] Update GetPodResources() to handle in-place pod resource updates
Implement initial support for the upcoming Kubernetes InPlacePodVerticalScaling
feature when doing Pod resource calculations. If present, the allocated
resources of each container will be used in preference to the requested
resources. Additionally, if the Pod status flag Resize is set to Proposed,
use the larger of both requested and allocated for each resource type.