[YUNIKORN-2630] Release context lock early for config changes
Release the lock of the context in the shim when processing is done.
When the config changes are sent to the core the k8shim should not be
locked. The context changes have been finalised at that point.

The core handles its own locking and serialises config changes that come
in from the k8shim.